+++ Bluepill-cli/Bluepill-cli/Simulator/SimulatorMonitor.m
@@ -15,6 +15,7 @@
 
 typedef NS_ENUM(NSInteger, SimulatorState) {
     Idle,
+    AppLaunched,
     Running,
     Completed
 };
@@ -172,6 +173,10 @@ typedef NS_ENUM(NSInteger, SimulatorState) {
 - (void)onOutputReceived:(NSString *)output {
     NSDate *currentTime = [NSDate date];
 
+    if (self.simulatorState == Idle) {
+        self.simulatorState = AppLaunched;
+    }
+
     self.currentOutputId++; // Increment the Output ID for this instance since we've moved on to the next bit of output
 
     __block NSUInteger previousOutputId = self.currentOutputId;
@@ -186,20 +191,24 @@ typedef NS_ENUM(NSInteger, SimulatorState) {
                               forTestName:(__self.currentTestName ?: __self.previousTestName)
                                   inClass:(__self.currentClassName ?: __self.previousClassName)];
         __self.exitStatus = BPExitStatusAppCrashed;
-        [[BPStats sharedStats] addSimulatorCrash];
+        [[BPStats sharedStats] addApplicationCrash];
     }
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(__self.maxTimeWithNoOutput * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
-        if (__self.currentOutputId == previousOutputId && __self.simulatorState == Running) {
+        if (__self.currentOutputId == previousOutputId && (__self.simulatorState == Running || __self.simulatorState == AppLaunched)) {
             NSString *testClass = (__self.currentClassName ?: __self.previousClassName);
             NSString *testName = (__self.currentTestName ?: __self.previousTestName);
-            [BPUtils printInfo:TIMEOUT withString:@" %10.6fs waiting for output from %@/%@",
-                                                   __self.maxTimeWithNoOutput, testClass, testName];
+            if (testClass == nil && testName == nil && (__self.simulatorState == AppLaunched || __self.simulatorState == Idle)) {
+                [BPUtils printInfo:ERROR withString:@"It appears that tests have not yet started. The test app has frozen prior to the first test."];
+            } else {
+                [BPUtils printInfo:TIMEOUT withString:@" %10.6fs waiting for output from %@/%@",
+                 __self.maxTimeWithNoOutput, testClass, testName];
+                [[BPStats sharedStats] endTimer:[NSString stringWithFormat:TEST_CASE_FORMAT, [BPStats sharedStats].attemptNumber, testClass, testName]];
+            }
             [__self stopTestsWithErrorMessage:@"Timed out waiting for the test to produce output. Test was aboorted."
                                   forTestName:testName
                                       inClass:testClass];
             __self.exitStatus = BPExitStatusTestTimeout;
-            [[BPStats sharedStats] endTimer:[NSString stringWithFormat:TEST_CASE_FORMAT, [BPStats sharedStats].attemptNumber, testClass, testName]];
-            [[BPStats sharedStats] addTestBPExitStatusTestTimeout];
+            [[BPStats sharedStats] addTestOutputTimeout];
         }
     });
     self.lastOutput = currentTime;
@@ -210,15 +219,12 @@ typedef NS_ENUM(NSInteger, SimulatorState) {
     // Timeout or crash on a test means we should skip it when we rerun the tests
     [self updateExecutedTestCaseList:testName inClass:testClass];
 
-    if (![[self.device stateString] isEqualToString:@"Shutdown"]) {
-        // self.appPID can be zero when running the parsing tests
-        // since we're not actually creating a simulator and running an app.
+    if (![[self.device stateString] isEqualToString:@"Shutdown"] && !self.config.testing_NoAppWillRun) {
         [BPUtils printInfo:ERROR withString:@"Will kill the process with appPID: %d", self.appPID];
-        if (self.appPID && (kill(self.appPID, 0) == 0) && (kill(self.appPID, SIGTERM) < 0)) {
-            [BPUtils printInfo:ERROR withString:@"Failed to kill the process with appPID: %d", self.appPID];
-            perror("kill");
-        } else {
-            [BPUtils printInfo:ERROR withString:@"Success killing the process with appPID: %d", self.appPID];
+        NSAssert(self.appPID > 0, @"Failed to find a valid PID");
+        if ((kill(self.appPID, 0) == 0) && (kill(self.appPID, SIGTERM) < 0)) {
+            [BPUtils printInfo:ERROR withString:@"Failed to kill the process with appPID: %d: %s",
+                self.appPID, strerror(errno)];
         }
     }
